fix: resolve all 272 mypy type errors (0 remaining)
Main patterns fixed:
- Add `| None` to parameters with `= None` default (implicit Optional)
- Fix _learn_with_shrinkage return type: dict[str, dict] → dict (keys are
  tuples in practice, not strings)
- Explicit Anthropic() kwargs instead of **dict unpacking (auto_feedback.py)
- Flask view functions: remove Response annotation (tuple[Response, int]
  is valid but mypy rejects it without ResponseReturnValue)
- str() cast on LLM response content (claude_advisor, baichuan_advisor)
- type: ignore[no-any-return] for json.loads() return sites
- Add trailing `return False` in fetch_and_log for exhausted retry loop
- Remove now-unused `flask.Response` import after annotation removal

+++ src/superhealth/dashboard/prediction/hyperlipidemia_risk.py
@@ -310,7 +310,7 @@ def _compute_score(
 
     # LDL-C 未达标加分
     if ldl_status == "未达标":
-        base += min(20, ldl_excess * 10)
+        base += int(min(20, ldl_excess * 10))
 
     # TG 加分
     base += tg_grade * 5
